third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_10.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = '<span style="direction: ltr">ipsum dolor sit</span> amet';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('ltr', 1, 'right', '<^s|pan style="direction: ltr">ipsum dolor sit amet');
testIt('ltr', 2, 'right', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('ltr', 3, 'right', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('ltr', 4, 'right', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('ltr', 5, 'right', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('ltr', 6, 'right', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'right', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 8, 'right', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 9, 'right', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 10, 'right', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 11, 'right', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 12, 'right', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'right', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'right', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'right', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'right', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'right', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'right', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'right', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'right', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'right', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'right', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'right', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'right', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'right', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'right', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'right', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'right', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'right', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('ltr', 30, 'right', '<span style="direction: ltr">i^p|sum dolor sit</span> amet');
testIt('ltr', 31, 'right', '<span style="direction: ltr">ip^s|um dolor sit</span> amet');
testIt('ltr', 32, 'right', '<span style="direction: ltr">ips^u|m dolor sit</span> amet');
testIt('ltr', 33, 'right', '<span style="direction: ltr">ipsu^m| dolor sit</span> amet');
testIt('ltr', 34, 'right', '<span style="direction: ltr">ipsum^ |dolor sit</span> amet');
testIt('ltr', 35, 'right', '<span style="direction: ltr">ipsum ^d|olor sit</span> amet');
testIt('ltr', 36, 'right', '<span style="direction: ltr">ipsum d^o|lor sit</span> amet');
testIt('ltr', 37, 'right', '<span style="direction: ltr">ipsum do^l|or sit</span> amet');
testIt('ltr', 38, 'right', '<span style="direction: ltr">ipsum dol^o|r sit</span> amet');
testIt('ltr', 39, 'right', '<span style="direction: ltr">ipsum dolo^r| sit</span> amet');
testIt('ltr', 40, 'right', '<span style="direction: ltr">ipsum dolor^ |sit</span> amet');
testIt('ltr', 41, 'right', '<span style="direction: ltr">ipsum dolor ^s|it</span> amet');
testIt('ltr', 42, 'right', '<span style="direction: ltr">ipsum dolor s^i|t</span> amet');
testIt('ltr', 43, 'right', '<span style="direction: ltr">ipsum dolor si^t|</span> amet');
testIt('ltr', 44, 'right', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('ltr', 45, 'right', '<span style="direction: ltr">ipsum dolor sit<^/|span> amet</span>');
testIt('ltr', 46, 'right', '<span style="direction: ltr">ipsum dolor sitspan^ |amet</span>');
testIt('ltr', 47, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 48, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 49, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 50, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 51, 'right', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('ltr', 52, 'right', '<span style="direction: ltr">ipsum dolor sit</span> ^a|met');
testIt('ltr', 53, 'right', '<span style="direction: ltr">ipsum dolor sit</span> a^m|et');
testIt('ltr', 54, 'right', '<span style="direction: ltr">ipsum dolor sit</span> am^e|t');
testIt('ltr', 55, 'right', '<span style="direction: ltr">ipsum dolor sit</span> ame^t|');

// ltr left character
testIt('ltr', 0, 'left', '|<span style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 1, 'left', '|<^span style="direction: ltr">ipsum dolor sit amet');
testIt('ltr', 2, 'left', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('ltr', 3, 'left', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('ltr', 4, 'left', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('ltr', 5, 'left', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('ltr', 6, 'left', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'left', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 8, 'left', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 9, 'left', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 10, 'left', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 11, 'left', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 12, 'left', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'left', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'left', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'left', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'left', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'left', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'left', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'left', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'left', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'left', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'left', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'left', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'left', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'left', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'left', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'left', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'left', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'left', '<span style="direction: ltr">|ipsum dolor sit</span> amet');
testIt('ltr', 30, 'left', '<span style="direction: ltr">|i^psum dolor sit</span> amet');
testIt('ltr', 31, 'left', '<span style="direction: ltr">i|p^sum dolor sit</span> amet');
testIt('ltr', 32, 'left', '<span style="direction: ltr">ip|s^um dolor sit</span> amet');
testIt('ltr', 33, 'left', '<span style="direction: ltr">ips|u^m dolor sit</span> amet');
testIt('ltr', 34, 'left', '<span style="direction: ltr">ipsu|m^ dolor sit</span> amet');
testIt('ltr', 35, 'left', '<span style="direction: ltr">ipsum| ^dolor sit</span> amet');
testIt('ltr', 36, 'left', '<span style="direction: ltr">ipsum |d^olor sit</span> amet');
testIt('ltr', 37, 'left', '<span style="direction: ltr">ipsum d|o^lor sit</span> amet');
testIt('ltr', 38, 'left', '<span style="direction: ltr">ipsum do|l^or sit</span> amet');
testIt('ltr', 39, 'left', '<span style="direction: ltr">ipsum dol|o^r sit</span> amet');
testIt('ltr', 40, 'left', '<span style="direction: ltr">ipsum dolo|r^ sit</span> amet');
testIt('ltr', 41, 'left', '<span style="direction: ltr">ipsum dolor| ^sit</span> amet');
testIt('ltr', 42, 'left', '<span style="direction: ltr">ipsum dolor |s^it</span> amet');
testIt('ltr', 43, 'left', '<span style="direction: ltr">ipsum dolor s|i^t</span> amet');
testIt('ltr', 44, 'left', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('ltr', 45, 'left', '<span style="direction: ltr">ipsum dolor sit|<^/span> amet</span>');
testIt('ltr', 46, 'left', '<span style="direction: ltr">ipsum dolor si|t^span amet</span>');
testIt('ltr', 47, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 48, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 49, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 50, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 51, 'left', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('ltr', 52, 'left', '<span style="direction: ltr">ipsum dolor sit</span>| ^amet');
testIt('ltr', 53, 'left', '<span style="direction: ltr">ipsum dolor sit</span> |a^met');
testIt('ltr', 54, 'left', '<span style="direction: ltr">ipsum dolor sit</span> a|m^et');
testIt('ltr', 55, 'left', '<span style="direction: ltr">ipsum dolor sit</span> am|e^t');

// ltr forward character
testIt('ltr', 0, 'forward', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('ltr', 1, 'forward', '<^s|pan style="direction: ltr">ipsum dolor sit amet');
testIt('ltr', 2, 'forward', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('ltr', 3, 'forward', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('ltr', 4, 'forward', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('ltr', 5, 'forward', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('ltr', 6, 'forward', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'forward', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 8, 'forward', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 9, 'forward', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 10, 'forward', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 11, 'forward', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 12, 'forward', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'forward', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'forward', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'forward', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'forward', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'forward', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'forward', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'forward', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'forward', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'forward', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'forward', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'forward', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'forward', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'forward', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'forward', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'forward', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'forward', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'forward', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('ltr', 30, 'forward', '<span style="direction: ltr">i^p|sum dolor sit</span> amet');
testIt('ltr', 31, 'forward', '<span style="direction: ltr">ip^s|um dolor sit</span> amet');
testIt('ltr', 32, 'forward', '<span style="direction: ltr">ips^u|m dolor sit</span> amet');
testIt('ltr', 33, 'forward', '<span style="direction: ltr">ipsu^m| dolor sit</span> amet');
testIt('ltr', 34, 'forward', '<span style="direction: ltr">ipsum^ |dolor sit</span> amet');
testIt('ltr', 35, 'forward', '<span style="direction: ltr">ipsum ^d|olor sit</span> amet');
testIt('ltr', 36, 'forward', '<span style="direction: ltr">ipsum d^o|lor sit</span> amet');
testIt('ltr', 37, 'forward', '<span style="direction: ltr">ipsum do^l|or sit</span> amet');
testIt('ltr', 38, 'forward', '<span style="direction: ltr">ipsum dol^o|r sit</span> amet');
testIt('ltr', 39, 'forward', '<span style="direction: ltr">ipsum dolo^r| sit</span> amet');
testIt('ltr', 40, 'forward', '<span style="direction: ltr">ipsum dolor^ |sit</span> amet');
testIt('ltr', 41, 'forward', '<span style="direction: ltr">ipsum dolor ^s|it</span> amet');
testIt('ltr', 42, 'forward', '<span style="direction: ltr">ipsum dolor s^i|t</span> amet');
testIt('ltr', 43, 'forward', '<span style="direction: ltr">ipsum dolor si^t|</span> amet');
testIt('ltr', 44, 'forward', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('ltr', 45, 'forward', '<span style="direction: ltr">ipsum dolor sit<^/|span> amet</span>');
testIt('ltr', 46, 'forward', '<span style="direction: ltr">ipsum dolor sitspan^ |amet</span>');
testIt('ltr', 47, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 48, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 49, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 50, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 51, 'forward', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('ltr', 52, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> ^a|met');
testIt('ltr', 53, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> a^m|et');
testIt('ltr', 54, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> am^e|t');
testIt('ltr', 55, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> ame^t|');

// ltr backward character
testIt('ltr', 0, 'backward', '|<span style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 1, 'backward', '|<^span style="direction: ltr">ipsum dolor sit amet');
testIt('ltr', 2, 'backward', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('ltr', 3, 'backward', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('ltr', 4, 'backward', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('ltr', 5, 'backward', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('ltr', 6, 'backward', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'backward', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 8, 'backward', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 9, 'backward', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 10, 'backward', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 11, 'backward', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 12, 'backward', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'backward', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'backward', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'backward', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'backward', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'backward', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'backward', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'backward', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'backward', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'backward', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'backward', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'backward', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'backward', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'backward', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'backward', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'backward', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'backward', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'backward', '<span style="direction: ltr">|ipsum dolor sit</span> amet');
testIt('ltr', 30, 'backward', '<span style="direction: ltr">|i^psum dolor sit</span> amet');
testIt('ltr', 31, 'backward', '<span style="direction: ltr">i|p^sum dolor sit</span> amet');
testIt('ltr', 32, 'backward', '<span style="direction: ltr">ip|s^um dolor sit</span> amet');
testIt('ltr', 33, 'backward', '<span style="direction: ltr">ips|u^m dolor sit</span> amet');
testIt('ltr', 34, 'backward', '<span style="direction: ltr">ipsu|m^ dolor sit</span> amet');
testIt('ltr', 35, 'backward', '<span style="direction: ltr">ipsum| ^dolor sit</span> amet');
testIt('ltr', 36, 'backward', '<span style="direction: ltr">ipsum |d^olor sit</span> amet');
testIt('ltr', 37, 'backward', '<span style="direction: ltr">ipsum d|o^lor sit</span> amet');
testIt('ltr', 38, 'backward', '<span style="direction: ltr">ipsum do|l^or sit</span> amet');
testIt('ltr', 39, 'backward', '<span style="direction: ltr">ipsum dol|o^r sit</span> amet');
testIt('ltr', 40, 'backward', '<span style="direction: ltr">ipsum dolo|r^ sit</span> amet');
testIt('ltr', 41, 'backward', '<span style="direction: ltr">ipsum dolor| ^sit</span> amet');
testIt('ltr', 42, 'backward', '<span style="direction: ltr">ipsum dolor |s^it</span> amet');
testIt('ltr', 43, 'backward', '<span style="direction: ltr">ipsum dolor s|i^t</span> amet');
testIt('ltr', 44, 'backward', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('ltr', 45, 'backward', '<span style="direction: ltr">ipsum dolor sit|<^/span> amet</span>');
testIt('ltr', 46, 'backward', '<span style="direction: ltr">ipsum dolor si|t^span amet</span>');
testIt('ltr', 47, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 48, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 49, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 50, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('ltr', 51, 'backward', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('ltr', 52, 'backward', '<span style="direction: ltr">ipsum dolor sit</span>| ^amet');
testIt('ltr', 53, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> |a^met');
testIt('ltr', 54, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> a|m^et');
testIt('ltr', 55, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> am|e^t');

// rtl right character
testIt('rtl', 0, 'right', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('rtl', 1, 'right', '<^s|pan style="direction: ltr">ipsum dolor sit amet');
testIt('rtl', 2, 'right', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('rtl', 3, 'right', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('rtl', 4, 'right', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('rtl', 5, 'right', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('rtl', 6, 'right', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'right', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 8, 'right', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 9, 'right', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 10, 'right', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 11, 'right', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 12, 'right', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'right', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'right', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'right', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'right', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'right', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'right', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'right', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'right', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'right', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'right', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'right', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'right', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'right', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'right', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'right', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'right', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'right', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('rtl', 30, 'right', '<span style="direction: ltr">i^p|sum dolor sit</span> amet');
testIt('rtl', 31, 'right', '<span style="direction: ltr">ip^s|um dolor sit</span> amet');
testIt('rtl', 32, 'right', '<span style="direction: ltr">ips^u|m dolor sit</span> amet');
testIt('rtl', 33, 'right', '<span style="direction: ltr">ipsu^m| dolor sit</span> amet');
testIt('rtl', 34, 'right', '<span style="direction: ltr">ipsum^ |dolor sit</span> amet');
testIt('rtl', 35, 'right', '<span style="direction: ltr">ipsum ^d|olor sit</span> amet');
testIt('rtl', 36, 'right', '<span style="direction: ltr">ipsum d^o|lor sit</span> amet');
testIt('rtl', 37, 'right', '<span style="direction: ltr">ipsum do^l|or sit</span> amet');
testIt('rtl', 38, 'right', '<span style="direction: ltr">ipsum dol^o|r sit</span> amet');
testIt('rtl', 39, 'right', '<span style="direction: ltr">ipsum dolo^r| sit</span> amet');
testIt('rtl', 40, 'right', '<span style="direction: ltr">ipsum dolor^ |sit</span> amet');
testIt('rtl', 41, 'right', '<span style="direction: ltr">ipsum dolor ^s|it</span> amet');
testIt('rtl', 42, 'right', '<span style="direction: ltr">ipsum dolor s^i|t</span> amet');
testIt('rtl', 43, 'right', '<span style="direction: ltr">ipsum dolor si^t|</span> amet');
testIt('rtl', 44, 'right', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('rtl', 45, 'right', '<span style="direction: ltr">ipsum dolor sit<^/|span> amet</span>');
testIt('rtl', 46, 'right', '<span style="direction: ltr">ipsum dolor sitspan^ |amet</span>');
testIt('rtl', 47, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 48, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 49, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 50, 'right', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 51, 'right', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('rtl', 52, 'right', '<span style="direction: ltr">ipsum dolor sit</span> ^a|met');
testIt('rtl', 53, 'right', '<span style="direction: ltr">ipsum dolor sit</span> a^m|et');
testIt('rtl', 54, 'right', '<span style="direction: ltr">ipsum dolor sit</span> am^e|t');
testIt('rtl', 55, 'right', '<span style="direction: ltr">ipsum dolor sit</span> ame^t|');

// rtl left character
testIt('rtl', 0, 'left', '|<span style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 1, 'left', '|<^span style="direction: ltr">ipsum dolor sit amet');
testIt('rtl', 2, 'left', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('rtl', 3, 'left', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('rtl', 4, 'left', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('rtl', 5, 'left', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('rtl', 6, 'left', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'left', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 8, 'left', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 9, 'left', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 10, 'left', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 11, 'left', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 12, 'left', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'left', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'left', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'left', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'left', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'left', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'left', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'left', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'left', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'left', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'left', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'left', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'left', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'left', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'left', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'left', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'left', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'left', '<span style="direction: ltr">|ipsum dolor sit</span> amet');
testIt('rtl', 30, 'left', '<span style="direction: ltr">|i^psum dolor sit</span> amet');
testIt('rtl', 31, 'left', '<span style="direction: ltr">i|p^sum dolor sit</span> amet');
testIt('rtl', 32, 'left', '<span style="direction: ltr">ip|s^um dolor sit</span> amet');
testIt('rtl', 33, 'left', '<span style="direction: ltr">ips|u^m dolor sit</span> amet');
testIt('rtl', 34, 'left', '<span style="direction: ltr">ipsu|m^ dolor sit</span> amet');
testIt('rtl', 35, 'left', '<span style="direction: ltr">ipsum| ^dolor sit</span> amet');
testIt('rtl', 36, 'left', '<span style="direction: ltr">ipsum |d^olor sit</span> amet');
testIt('rtl', 37, 'left', '<span style="direction: ltr">ipsum d|o^lor sit</span> amet');
testIt('rtl', 38, 'left', '<span style="direction: ltr">ipsum do|l^or sit</span> amet');
testIt('rtl', 39, 'left', '<span style="direction: ltr">ipsum dol|o^r sit</span> amet');
testIt('rtl', 40, 'left', '<span style="direction: ltr">ipsum dolo|r^ sit</span> amet');
testIt('rtl', 41, 'left', '<span style="direction: ltr">ipsum dolor| ^sit</span> amet');
testIt('rtl', 42, 'left', '<span style="direction: ltr">ipsum dolor |s^it</span> amet');
testIt('rtl', 43, 'left', '<span style="direction: ltr">ipsum dolor s|i^t</span> amet');
testIt('rtl', 44, 'left', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('rtl', 45, 'left', '<span style="direction: ltr">ipsum dolor sit|<^/span> amet</span>');
testIt('rtl', 46, 'left', '<span style="direction: ltr">ipsum dolor si|t^span amet</span>');
testIt('rtl', 47, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 48, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 49, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 50, 'left', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 51, 'left', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('rtl', 52, 'left', '<span style="direction: ltr">ipsum dolor sit</span>| ^amet');
testIt('rtl', 53, 'left', '<span style="direction: ltr">ipsum dolor sit</span> |a^met');
testIt('rtl', 54, 'left', '<span style="direction: ltr">ipsum dolor sit</span> a|m^et');
testIt('rtl', 55, 'left', '<span style="direction: ltr">ipsum dolor sit</span> am|e^t');

// rtl forward character
testIt('rtl', 0, 'forward', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('rtl', 1, 'forward', '<^s|pan style="direction: ltr">ipsum dolor sit amet');
testIt('rtl', 2, 'forward', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('rtl', 3, 'forward', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('rtl', 4, 'forward', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('rtl', 5, 'forward', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('rtl', 6, 'forward', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'forward', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 8, 'forward', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 9, 'forward', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 10, 'forward', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 11, 'forward', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 12, 'forward', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'forward', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'forward', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'forward', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'forward', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'forward', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'forward', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'forward', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'forward', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'forward', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'forward', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'forward', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'forward', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'forward', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'forward', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'forward', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'forward', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'forward', '<span style="direction: ltr">^i|psum dolor sit</span> amet');
testIt('rtl', 30, 'forward', '<span style="direction: ltr">i^p|sum dolor sit</span> amet');
testIt('rtl', 31, 'forward', '<span style="direction: ltr">ip^s|um dolor sit</span> amet');
testIt('rtl', 32, 'forward', '<span style="direction: ltr">ips^u|m dolor sit</span> amet');
testIt('rtl', 33, 'forward', '<span style="direction: ltr">ipsu^m| dolor sit</span> amet');
testIt('rtl', 34, 'forward', '<span style="direction: ltr">ipsum^ |dolor sit</span> amet');
testIt('rtl', 35, 'forward', '<span style="direction: ltr">ipsum ^d|olor sit</span> amet');
testIt('rtl', 36, 'forward', '<span style="direction: ltr">ipsum d^o|lor sit</span> amet');
testIt('rtl', 37, 'forward', '<span style="direction: ltr">ipsum do^l|or sit</span> amet');
testIt('rtl', 38, 'forward', '<span style="direction: ltr">ipsum dol^o|r sit</span> amet');
testIt('rtl', 39, 'forward', '<span style="direction: ltr">ipsum dolo^r| sit</span> amet');
testIt('rtl', 40, 'forward', '<span style="direction: ltr">ipsum dolor^ |sit</span> amet');
testIt('rtl', 41, 'forward', '<span style="direction: ltr">ipsum dolor ^s|it</span> amet');
testIt('rtl', 42, 'forward', '<span style="direction: ltr">ipsum dolor s^i|t</span> amet');
testIt('rtl', 43, 'forward', '<span style="direction: ltr">ipsum dolor si^t|</span> amet');
testIt('rtl', 44, 'forward', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('rtl', 45, 'forward', '<span style="direction: ltr">ipsum dolor sit<^/|span> amet</span>');
testIt('rtl', 46, 'forward', '<span style="direction: ltr">ipsum dolor sitspan^ |amet</span>');
testIt('rtl', 47, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 48, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 49, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 50, 'forward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 51, 'forward', '<span style="direction: ltr">ipsum dolor sit</span>^ |amet');
testIt('rtl', 52, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> ^a|met');
testIt('rtl', 53, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> a^m|et');
testIt('rtl', 54, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> am^e|t');
testIt('rtl', 55, 'forward', '<span style="direction: ltr">ipsum dolor sit</span> ame^t|');

// rtl backward character
testIt('rtl', 0, 'backward', '|<span style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 1, 'backward', '|<^span style="direction: ltr">ipsum dolor sit amet');
testIt('rtl', 2, 'backward', '<s|pan style="direction: ltr">ipsum dolor sit amet</s|pan>');
testIt('rtl', 3, 'backward', '<sp|an style="direction: ltr">ipsum dolor sit amet</sp|an>');
testIt('rtl', 4, 'backward', '<spa|n style="direction: ltr">ipsum dolor sit amet</spa|n>');
testIt('rtl', 5, 'backward', '<span| style="direction: ltr">ipsum dolor sit amet</span|>');
testIt('rtl', 6, 'backward', '<span |style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'backward', '<span s|tyle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 8, 'backward', '<span st|yle="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 9, 'backward', '<span sty|le="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 10, 'backward', '<span styl|e="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 11, 'backward', '<span style|="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 12, 'backward', '<span ltr" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'backward', '<span style="|direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'backward', '<span style="d|irection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'backward', '<span style="di|rection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'backward', '<span style="dir|ection: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'backward', '<span style="dire|ction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'backward', '<span style="direc|tion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'backward', '<span style="direct|ion: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'backward', '<span style="directi|on: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'backward', '<span style="directio|n: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'backward', '<span style="direction|: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'backward', '<span style="direction:| ltr">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'backward', '<span style="direction: |ltr">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'backward', '<span style="direction: l|tr">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'backward', '<span style="direction: lt|r">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'backward', '<span style="direction: ltr|">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'backward', '<span | style="direction: ltr">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'backward', '<span style="direction: ltr">|ipsum dolor sit</span> amet');
testIt('rtl', 30, 'backward', '<span style="direction: ltr">|i^psum dolor sit</span> amet');
testIt('rtl', 31, 'backward', '<span style="direction: ltr">i|p^sum dolor sit</span> amet');
testIt('rtl', 32, 'backward', '<span style="direction: ltr">ip|s^um dolor sit</span> amet');
testIt('rtl', 33, 'backward', '<span style="direction: ltr">ips|u^m dolor sit</span> amet');
testIt('rtl', 34, 'backward', '<span style="direction: ltr">ipsu|m^ dolor sit</span> amet');
testIt('rtl', 35, 'backward', '<span style="direction: ltr">ipsum| ^dolor sit</span> amet');
testIt('rtl', 36, 'backward', '<span style="direction: ltr">ipsum |d^olor sit</span> amet');
testIt('rtl', 37, 'backward', '<span style="direction: ltr">ipsum d|o^lor sit</span> amet');
testIt('rtl', 38, 'backward', '<span style="direction: ltr">ipsum do|l^or sit</span> amet');
testIt('rtl', 39, 'backward', '<span style="direction: ltr">ipsum dol|o^r sit</span> amet');
testIt('rtl', 40, 'backward', '<span style="direction: ltr">ipsum dolo|r^ sit</span> amet');
testIt('rtl', 41, 'backward', '<span style="direction: ltr">ipsum dolor| ^sit</span> amet');
testIt('rtl', 42, 'backward', '<span style="direction: ltr">ipsum dolor |s^it</span> amet');
testIt('rtl', 43, 'backward', '<span style="direction: ltr">ipsum dolor s|i^t</span> amet');
testIt('rtl', 44, 'backward', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('rtl', 45, 'backward', '<span style="direction: ltr">ipsum dolor sit|<^/span> amet</span>');
testIt('rtl', 46, 'backward', '<span style="direction: ltr">ipsum dolor si|t^span amet</span>');
testIt('rtl', 47, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 48, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 49, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 50, 'backward', '<span style="direction: ltr">ipsum dolor sit amet</span>');
testIt('rtl', 51, 'backward', '<span style="direction: ltr">ipsum dolor si|t^</span> amet');
testIt('rtl', 52, 'backward', '<span style="direction: ltr">ipsum dolor sit</span>| ^amet');
testIt('rtl', 53, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> |a^met');
testIt('rtl', 54, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> a|m^et');
testIt('rtl', 55, 'backward', '<span style="direction: ltr">ipsum dolor sit</span> am|e^t');
</script>
